requirements.txt
altgraph==0.17.4
anyio==4.3.0
auto-py-to-exe==2.42.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
Eel==0.16.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
packaging==24.0
pefile==2023.2.7
pycparser==2.21
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyparsing==3.1.2
python-telegram-bot==21.0.1
pywin32-ctypes==0.2.2
sniffio==1.3.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.2
